+++ src/Git_Updater/Bootstrap.php
@@ -21,14 +21,6 @@ if ( ! defined( 'WPINC' ) ) {
 	die;
 }
 
-// Load textdomain.
-add_action(
-	'init',
-	function () {
-		load_plugin_textdomain( 'git-updater' );
-	}
-);
-
 /**
  * Class Bootstrap
  */