+++ src/Git_Updater/Settings.php
@@ -306,8 +306,8 @@ class Settings {
 		<div class="wrap git-updater-settings">
 			<h1>
 				<a href="https://github.com/afragen/git-updater" target="_blank"><img src="<?php echo esc_attr( $logo ); ?>" alt="Git Updater logo" /></a><br>
-				<?php esc_html_e( __( 'Git Updater', 'git-updater' ) ); ?>
-				<span class="description"><?php esc_html_e( ' v' . $this->get_plugin_version() ); ?></span>
+				<?php esc_html_e( 'Git Updater', 'git-updater' ); ?>
+				<span class="description"><?php echo esc_html( ' v' . $this->get_plugin_version() ); ?></span>
 			</h1>
 			<?php $this->options_tabs(); ?>
 			<?php $this->admin_page_notices(); ?>