+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* remove `load_plugin_textdomain()`
 
 #### 12.6.0 / 2024-10-13
 * check existence of `FS__RESOLVE_CLONE_AS` before setting