+++ CHANGES.md
@@ -1,6 +1,7 @@
 #### [unreleased]
 * remove `load_plugin_textdomain()`
 * add git host icon to single site theme description
+* don't save to GitHub.com access token from single repo remote install
 
 #### 12.6.0 / 2024-10-13
 * check existence of `FS__RESOLVE_CLONE_AS` before setting